requirements/common.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==7.2.0
pytz==2021.1
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
